fix: code changes with Nelson

+++ src/com/leo/booking/CarBookingService.java
@@ -4,6 +4,8 @@ import com.leo.car.Car;
 import com.leo.car.CarService;
 import com.leo.user.User;
 import com.leo.user.UserService;
+
+import java.io.FileWriter;
 import java.util.UUID;
 import java.time.LocalDateTime;
 import java.time.LocalDate;
@@ -29,17 +31,32 @@ public class CarBookingService {
     }
 
     public void deleteBooking(UUID bookingId) {
+        // does bookign exitsts if not throw
+        if (findBookingById(bookingId) == null) {
+            throw new IllegalStateException("");
+        }
         bookingDAO.deleteBooking(bookingId);
     }
 
     public CarBooking bookCar(UUID userId, UUID carId, LocalDate startDate, LocalDate endDate) {
         User user = userService.findUserById(userId);
+        // is null
         Car car = carService.findCarById(carId);
+        // is null throw
+
+        // is car being rented
+        // throw car in use execption
+        for (CarBooking booking : bookingDAO.getBookings()) {
+            if(booking.getCar().getId().equals(carId)){
+                throw new IllegalStateException("car being rented");
+            }
+        }
 
         long days = ChronoUnit.DAYS.between(startDate, endDate);
         if(days <= 0) {
-            throw new RuntimeException("End date must be after start date");
+            throw new IllegalStateException("End date must be after start date");
         }
+
 
         double totalPrice = car.getRentalPricePerDay() * days;
         UUID bookingId = UUID.randomUUID();
